split main file
